+++ java-oop-3/src/Main.java
@@ -1,8 +1,8 @@
-import controller.Controller;
+import controller.CharacterController;
 
 public class Main {
     public static void main(String[] args) {
-        Controller controller = new Controller();
-        controller.run();
+        CharacterController characterController = new CharacterController();
+        characterController.run();
     }
 }
