+++ java-oop-3/src/view/CharacterInfo.java
@@ -4,9 +4,10 @@ import model.Character;
 
 public class CharacterInfo {
     public void characterInfo(Character character) {
-        System.out.println();
-        System.out.println("캐릭터: " + character.getName());
-        System.out.println("체력: " + character.getHp());
-        System.out.println("공격력: " + character.getAttackPower());
+        System.out.printf("""
+                \n캐릭터: %s
+                체력: %d
+                공격력: %d
+                """, character.getName(), character.getHp(), character.getAttackPower());
     }
 }
